migrating to GW 2.0 refactoring

+++ src/test/java/org/mule/AbstractTemplateTest.java
@@ -12,7 +12,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Properties;
 
 import javax.ws.rs.client.Entity;
@@ -43,14 +42,15 @@ import org.xml.sax.SAXException;
  */
 public abstract class AbstractTemplateTest {
 
+	private static final String PROXY_URL = "http://localhost:8081";
+	private static final String HTTP_REQUEST = "http:request";
+	private static final String HTTP_REQUEST_CONFIG = "http:request-config";
 	protected static final String KEY_STORE_PASSWORD = "keyStorePassword";
-	protected static final String KEYSTORE_NAME = "keystore.jks";
-	private static final String META_INF_FOLDER = "META-INF/src/main/resources/";
+	protected static final String KEYSTORE_NAME = "mule.jks";
 	protected static final String EXPORT_FOLDER = "export";
 	private static final String INPUT_FOLDER = "input";
 	protected String IMPLEMENTATION_URI = "implementation.uri";
 	
-	protected static final String CONFIG_NAME = "config.properties";
 	protected static final String TMP_FOLDER = "tmp";
 	protected final HashMap<String, String> input = new HashMap<String, String>();
 	protected String apiNameId;
@@ -62,13 +62,17 @@ public abstract class AbstractTemplateTest {
 	private static String PROXY_URI;
 	private static String LOGIN_URI;
 	private static String DOWNLOAD_PROXY_URI;
-	private static final String HTTPS_PROTOCOL = "https";
+	private static final String HTTPS_PROTOCOL = "https";	
 	protected String MESSAGE = "HTTPS connection established";
 	protected String proxyAppZip;
 	protected static String GATEWAY_APPS_FOLDER;
 	protected static String GATEWAY_VERSION;
 	protected static Logger LOGGER = LogManager.getLogger(AbstractTemplateTest.class);
 	protected final String PROXY_URI_KEY = "proxy.uri";
+	protected final String PROXY_HOST_KEY = "implementation.host";
+	protected final String PROXY_PORT_KEY = "implementation.port";
+	protected final String PROXY_PATH_KEY = "implementation.path";
+	protected boolean TEST_WITH_GATEWAY;
 	
 	/**
 	 * prepares test data:
@@ -100,6 +104,7 @@ public abstract class AbstractTemplateTest {
     	LOGIN_URI = props.getProperty("loginUri");
     	DOWNLOAD_PROXY_URI = props.getProperty("downloadProxyUri");
     	GATEWAY_VERSION = props.getProperty("gatewayVersion");
+    	TEST_WITH_GATEWAY = Boolean.valueOf(props.getProperty("testWithGateway"));
     	LOGGER.info("Signing in to Anypoint Platform with a user:" + USER);
     	final ResteasyClient client = new ResteasyClientBuilder().build();
         final ResteasyWebTarget target = client.target(PROXY_URI + LOGIN_URI);
@@ -193,48 +198,29 @@ public abstract class AbstractTemplateTest {
 		
 		assertTrue("Modified zip file should be created", file.exists());
 		final String tmpFolder = TEST_RESOURCES_FOLDER + File.separator + TMP_FOLDER + File.separator;
-		FileUtils.deleteDirectory(new File(GATEWAY_APPS_FOLDER + file.getName().substring(0, file.getName().lastIndexOf("."))));
-		LOGGER.info("Deleting GW dir: " + GATEWAY_APPS_FOLDER + file.getName().substring(0, file.getName().lastIndexOf(".")));
-		FileUtils.copyFile(file, new File(GATEWAY_APPS_FOLDER + file.getName()));
+		if (TEST_WITH_GATEWAY){
+			FileUtils.copyFile(file, new File(GATEWAY_APPS_FOLDER + file.getName()));
+		}
+		else{
+			LOGGER.info("Skipping testing with Gateway..");
+		}
 		// need to wait till the API gateway starts a proxy app
 		Thread.sleep(15000);
 		Zipper.unZip(file, TEST_RESOURCES_FOLDER + File.separator + TMP_FOLDER);
 		
 		final Document proxy = readXML(tmpFolder + "proxy.xml");
-		final Element httpsConnector = (Element) proxy.getElementsByTagName("https:connector").item(0);		
-		assertNotNull("Proxy file should contain HTTPS connector element", httpsConnector);		
-		assertEquals("HTTPS connector element should contain TLS server element", "https:tls-server", httpsConnector.getChildNodes().item(1).getNodeName());
+		final Element httpRequestConfig = (Element) proxy.getElementsByTagName(HTTP_REQUEST_CONFIG).item(0);
+		assertNotNull("Proxy file should contain HTTP request config element", httpRequestConfig);		
+		assertNotNull("Proxy file should contain HTTPS connector element", httpRequestConfig.getAttribute("protocol").equals(HTTPS_PROTOCOL));		
 		
-		final Element httpsOutbound = (Element) proxy.getElementsByTagName("https:outbound-endpoint").item(0);
+		final Element httpsOutbound = (Element) proxy.getElementsByTagName(HTTP_REQUEST).item(0);
 		
 		assertNotNull("Proxy file should contain HTTPS outbound element", httpsOutbound);		
 		
 		assertTrue("Keystore should be packaged", new File(tmpFolder + KEYSTORE_NAME).exists());
-		
-		final List<String> config = FileUtils.readLines(new File(tmpFolder + File.separator + "classes" + File.separator + CONFIG_NAME));
-		final List<String> config1 = FileUtils.readLines(new File(tmpFolder + File.separator + META_INF_FOLDER + File.separator + CONFIG_NAME));
-		final String implUri = getImplementationUri(config);	
-		LOGGER.info("Implementation.uri: " + implUri);
-		assertTrue("'Classes' Implementation URI should start with https", implUri.contains(HTTPS_PROTOCOL));
-		final String implUri1 = getImplementationUri(config1);
-		assertTrue("'META-INF' Implementation URI should start with https", implUri1.contains(HTTPS_PROTOCOL));						
-        
+									       
 	}
 
-	/**
-	 * returns an implementation.uri configuration parameter
-	 * @param config a list of strings containing key-value pairs of configuration parameters 
-	 * @return value for IMPLEMENTATION_URI key
-	 */
-	private String getImplementationUri(final List<String> config) {
-		for (final String line : config){
-			if (line.startsWith(IMPLEMENTATION_URI)){				
-				return line;
-			}
-		}
-		return null;
-	}
-	
 	/**
 	 * reads an XML file and returns as org.w3c.dom.Document instance 
 	 * @param filepath path to XML file
@@ -278,13 +264,18 @@ public abstract class AbstractTemplateTest {
 	 * @throws IOException  thrown by FileUtils.copyFile
 	 */
 	protected void deployHTTPS() throws IOException{
-		LOGGER.info("Deploying HTTPS endpoint to a local gateway: " + GATEWAY_APPS_FOLDER);
-    	new File(GATEWAY_APPS_FOLDER + "https-test-anchor.txt").delete();
-    	FileUtils.copyFile(new File(TEST_RESOURCES_FOLDER + File.separator + "https-test.zip"), new File(GATEWAY_APPS_FOLDER + "https-test.zip"));
-    	try {
-			Thread.sleep(15000);
-		} catch (final InterruptedException e) {			
-			e.printStackTrace();
+		if (TEST_WITH_GATEWAY){
+			LOGGER.info("Deploying HTTPS endpoint to a local gateway: " + GATEWAY_APPS_FOLDER);
+	    	new File(GATEWAY_APPS_FOLDER + "https-test-anchor.txt").delete();
+	    	FileUtils.copyFile(new File(TEST_RESOURCES_FOLDER + File.separator + "https-test.zip"), new File(GATEWAY_APPS_FOLDER + "https-test.zip"));
+	    	try {
+				Thread.sleep(15000);
+			} catch (final InterruptedException e) {			
+				e.printStackTrace();
+			}
+		}
+		else{
+			LOGGER.info("Skipping deploying HTTPS endpoint to Gateway..");
 		}
 	}
 	
@@ -293,30 +284,39 @@ public abstract class AbstractTemplateTest {
 	 * @throws IOException thrown by FileUtils.copyFile
 	 */
 	protected void deployHTTPSforWSDL() throws IOException{
-		
-    	new File(GATEWAY_APPS_FOLDER + "xml-only-soap-web-service-anchor.txt").delete();
-    	FileUtils.copyFile(new File(TEST_RESOURCES_FOLDER + File.separator + "xml-only-soap-web-service.zip"), new File(GATEWAY_APPS_FOLDER + "xml-only-soap-web-service.zip"));
-    	try {
-			Thread.sleep(20000);
-		} catch (final InterruptedException e) {			
-			e.printStackTrace();
+		if (TEST_WITH_GATEWAY){
+	    	new File(GATEWAY_APPS_FOLDER + "xml-only-soap-web-service-anchor.txt").delete();
+	    	FileUtils.copyFile(new File(TEST_RESOURCES_FOLDER + File.separator + "xml-only-soap-web-service.zip"), new File(GATEWAY_APPS_FOLDER + "xml-only-soap-web-service.zip"));
+	    	try {
+				Thread.sleep(20000);
+			} catch (final InterruptedException e) {			
+				e.printStackTrace();
+			}
 		}
+    	else{
+			LOGGER.info("Skipping deploying HTTPS endpoint to Gateway..");
+		}		
 	}
-	
+		
 	/**
-	 * retrieves a proxy URI configuration parameter from the proxy app
-	 * @return proxy URI
-	 * @throws IOException thrown by FileUtils.readLines
+	 * makes a HTTP GET request using the proxy URI parameter
+	 * @return HTTP response body
+	 * @throws IllegalArgumentException thrown by ResteasyClient.target
+	 * @throws NullPointerException thrown by ResteasyClient.target
+	 * @throws IOException thrown by ResteasyClient.target
 	 */
-	protected String getProxyUriFromProperties() throws IOException{
-		final String path = TEST_RESOURCES_FOLDER + File.separator + TMP_FOLDER + File.separator + File.separator + "classes" + File.separator + CONFIG_NAME;
-    	
-		final List<String> lines = FileUtils.readLines(new File(path));
-		for (final String line : lines){
-			if (line.startsWith(PROXY_URI_KEY))
-				return line.substring(line.indexOf("=") + 1);
+	protected void makeTestRequest() throws IllegalArgumentException, NullPointerException, IOException{	
+		if (TEST_WITH_GATEWAY){
+			final ResteasyClient client = new ResteasyClientBuilder().build();
+	        final ResteasyWebTarget target = client.target(PROXY_URL);
+	        LOGGER.info("Making HTTP call: " + PROXY_URL);        
+	        final int response =  target.request().get().getStatus();
+	        LOGGER.info("HTTP response: " + response);
+			assertEquals("HTTPS request should be successful", 200, response);
 		}
-    	return null;    	
+		else{
+			LOGGER.info("Skipping testing with Gateway..");
+		}
 	}
 	
 	/**
@@ -326,20 +326,31 @@ public abstract class AbstractTemplateTest {
 	 * @throws NullPointerException thrown by ResteasyClient.target
 	 * @throws IOException thrown by ResteasyClient.target
 	 */
-	protected String makeTestRequest() throws IllegalArgumentException, NullPointerException, IOException{
-		final ResteasyClient client = new ResteasyClientBuilder().build();
-        final ResteasyWebTarget target = client.target(getProxyUriFromProperties());
-        LOGGER.info("Making HTTP call: " + getProxyUriFromProperties());
-        return target.request().get(String.class);               
+	protected void makeTestRequest(String path, String body) throws IllegalArgumentException, NullPointerException, IOException{				
+		if (TEST_WITH_GATEWAY){
+			final ResteasyClient client = new ResteasyClientBuilder().build();
+	        final ResteasyWebTarget target = client.target(PROXY_URL + path);
+	        LOGGER.info("Making HTTP call: " + PROXY_URL + path);        
+	        final Response response = target.request().post(Entity.entity(body, "text/plain"));
+	        LOGGER.info("HTTP response: " + response.getStatus());
+			assertEquals("HTTPS request should be successful", 200, response.getStatus());
+		}
+		else{
+			LOGGER.info("Skipping testing with Gateway..");
+		}
 	}
 	
-	
-	protected void testWsdlReqest() throws IllegalArgumentException, NullPointerException, IOException{
-		final ResteasyClient client = new ResteasyClientBuilder().build();
-        final ResteasyWebTarget target = client.target(getProxyUriFromProperties() + "?wsdl");
-        final Response response = target.request().get();    
-        LOGGER.info("Making HTTP call: " + target.getUri()); 
-        assertEquals("HTTPS request should be successful", 200, response.getStatus());
+	protected Properties initGatewayParams() {
+		final Properties props = new Properties();
+    	try {
+    		props.load(new FileInputStream(TEST_RESOURCES_FOLDER + File.separator + "test.properties"));
+    	} catch (final Exception e) {
+    		LOGGER.info("Error occured while reading test.properties" + e);
+    	} 
+    	TEST_WITH_GATEWAY = Boolean.valueOf(props.getProperty("testWithGateway"));
+    	GATEWAY_APPS_FOLDER = props.getProperty("gatewayAppDir");
+    	GATEWAY_VERSION = props.getProperty("gatewayVersion");
+		return props;
 	}
-	
+		
 }
